+++ EX4/src/services/CartService.java
@@ -3,7 +3,7 @@ package EX4.src.services;
 import EX4.src.models.Cart;
 
 public interface CartService {
-    void create(Integer userId);
-    Cart getById(Integer id);
-    void addProduct(Integer cartId, Integer productId, double price, int amount);
+    void create(Long userId);
+    Cart getById(Long id);
+    void addProduct(Long cartId, Long productId, Double price, Integer amount);
 }
